Huge Differences
I totally wrote the algorythm to calculate the whole stuff with a map in
Haskell and made Haskell and C# communicate with each other. And I wrote
a map editor for C#, so you can easily edit the maps for the simulating.
The next step is,to put all together :)